fix(demucs): type SeparationRequest.segment as int to match the CLI
demucs.separate defines --segment as type=int (verified against demucs
4.0.1). SeparateConfig and the API Form field already use int|None; align
the SeparationRequest dataclass, which was left float|None during the
upstream rebase.

+++ demucs_svc/separation/base.py
@@ -18,7 +18,7 @@ class SeparationRequest:
     requested_device: str
     output_format: str
     mp3_bitrate: int | None
-    segment: float | None = None
+    segment: int | None = None
     shifts: int | None = None
     jobs: int | None = None
     overlap: float | None = None
